rename paths for new locations
