+++ src/platform/silabs/wifi/SiWx/ncp/efx32_ncp_host.c
@@ -29,8 +29,8 @@
 #include "sl_si91x_status.h"
 #include "sl_status.h"
 #include "sl_wifi_constants.h"
-#include <platform/silabs/wifi/SiWx/ncp-platform/sl_board_configuration.h>
-#include <platform/silabs/wifi/ncp-common/spi_multiplex.h>
+#include <platform/silabs/wifi/SiWx/ncp/sl_board_configuration.h>
+#include <platform/silabs/wifi/ncp/spi_multiplex.h>
 #include <stdbool.h>
 #include <string.h>
