+++ src/platform/silabs/wifi/SiWx/ncp/sl_si91x_ncp_utility.h
@@ -29,8 +29,8 @@
 #include "silabs_utils.h"
 #include "sl_spidrv_exp_config.h"
 #include "sl_status.h"
-#include <platform/silabs/wifi/SiWx/ncp-platform/sl_board_configuration.h>
-#include <platform/silabs/wifi/ncp-common/spi_multiplex.h>
+#include <platform/silabs/wifi/SiWx/ncp/sl_board_configuration.h>
+#include <platform/silabs/wifi/ncp/spi_multiplex.h>
 
 #define USART_INITSYNC_BAUDRATE 12500000
